+++ src/tir/analysis/control_flow_graph.h
@@ -399,7 +399,7 @@ class ControlFlowGraph {
  public:
   /* \brief Extract the touch pattern from a TIR statement
    */
-  explicit ControlFlowGraph(const Stmt& stmt, int max_simplification_steps = 0,
+  explicit ControlFlowGraph(const Stmt& stmt, int64_t max_simplification_steps = 0,
                             size_t max_revisits = 5);
 
   /* \brief Check if a write is overwritten without impacting final results
@@ -658,7 +658,7 @@ class ControlFlowGraph {
   size_t max_revisits_;
 
   /*! \brief The maximum number of revisits while flowing constraints */
-  int max_simplification_steps_;
+  int64_t max_simplification_steps_;
 };
 
 }  // namespace tir
